remove unused links from types // bind favorite repositories container via extensionpoint
